+++ src/QuickMarcEditor/QuickMarcDeriveWrapper.js
@@ -84,6 +84,7 @@ const QuickMarcDeriveWrapper = ({
       marcRecord => autopopulateSubfieldSection(marcRecord, marcType),
       marcRecord => cleanBytesFields(marcRecord, fixedFieldSpec, marcType),
       marcRecord => formatLeaderForSubmit(marcType, marcRecord),
+      combineSplitFields,
     )(formValues);
 
     return formValuesForDerive;
@@ -101,10 +102,7 @@ const QuickMarcDeriveWrapper = ({
       return complete();
     }
 
-    const formValuesToProcess = flow(
-      prepareForSubmit,
-      combineSplitFields,
-    )(formValues);
+    const formValuesToProcess = prepareForSubmit(formValues);
 
     showCallout({ messageId: 'ui-quick-marc.record.saveNew.onSave' });